refactor(scripts): migrate validate-env CLI to shared lib.validation schema (ADS-707)
Moves the pure validateEnv() function into lib.validation/src/schemas/env.ts
so the CLI gate and the backend boot validator share one source of truth.
The boot-time validateEnvironment() wrapper (Winston logging + non-prod
"missing required" downgrade) stays in service.backend.

Replaces scripts/validate-env.mjs with scripts/validate-env.ts. Preserves
the CLI contract: colour output, exit codes, "X required vars missing"
summary line, --env-file/--example/--staging-env/--json flags. The new
script imports the schema via a relative source path so the CLI works in
CI before npm run build:libs has produced lib.validation/dist.

Adds tsx to root devDependencies and switches the validate:env npm script
to run via tsx. Updates doc references (.env.example, scripts/README.md,
service.backend/SECURITY.md, docs/operations/deploy.md) and the
check-env-example.mjs comment to point at the new .ts file.
