+++ .env.example
@@ -24,7 +24,7 @@
 #   * ENCRYPTION_KEY                                                   (auto-generated)
 #
 # A CI guard (`scripts/check-env-example.mjs`) keeps this list in sync with
-# `scripts/validate-env.mjs` — add new required vars there and re-run
+# `scripts/validate-env.ts` — add new required vars there and re-run
 # `npm run check:env-example` locally before pushing.
 #
 # =============================================================================
