+++ lib.validation/src/schemas/env.ts
@@ -1,16 +1,12 @@
 /**
- * Shared environment-variable schema (ADS-707).
+ * Shared environment-variable schema and validator (ADS-707).
  *
  * Single source of truth consumed by:
  *   - service.backend startup validator (validateEnvironment)
- *   - scripts/validate-env.mjs CLI gate (follow-up: migrate to tsx import)
+ *   - scripts/validate-env.ts CLI gate
  *
  * Pure schema + pure refinement functions — no logger / no side effects, so
  * the same module is safe to import from the boot path and the CLI tool.
- *
- * Note: this is the scaffold tranche of ADS-707. The .mjs CLI script still
- * carries its own rule copy and must be migrated in a follow-up so the two
- * paths cannot drift again. The backend validator is the first consumer.
  */
 import { z } from 'zod';
 
@@ -151,3 +147,203 @@ export const DISTINCT_SECRET_PAIRS: ReadonlyArray<readonly [string, string]> = [
   ['UPLOAD_SIGNING_SECRET', 'ENCRYPTION_KEY'],
   ['UPLOAD_SIGNING_SECRET', 'JWT_REPORT_SHARE_SECRET'],
 ] as const;
+
+// ---------------------------------------------------------------------------
+// Pure validateEnv — used by both the backend boot path and the CLI gate.
+// No logging / no process.exit; callers decide how to surface issues.
+// ---------------------------------------------------------------------------
+
+export type ValidationIssue = {
+  path: string;
+  message: string;
+  level: 'error' | 'warning';
+};
+
+export type EnvValidationResult = {
+  ok: boolean;
+  errors: ValidationIssue[];
+  warnings: ValidationIssue[];
+};
+
+type EnvMap = Record<string, string | undefined>;
+
+const distinctSecretsCheck = (env: EnvMap): ValidationIssue[] => {
+  const issues: ValidationIssue[] = [];
+  for (const [a, b] of DISTINCT_SECRET_PAIRS) {
+    const va = env[a];
+    const vb = env[b];
+    if (va && vb && va === vb) {
+      issues.push({
+        path: `${a}/${b}`,
+        message:
+          `${a} and ${b} must be distinct. Reusing secrets increases compromise blast radius. ` +
+          'Generate new secrets with: npm run secrets:generate',
+        level: 'error',
+      });
+    }
+  }
+  return issues;
+};
+
+const dbNameCheck = (env: EnvMap): ValidationIssue[] => {
+  const nodeEnv = env.NODE_ENV;
+  const issues: ValidationIssue[] = [];
+  // ADS-409: each environment requires its own DB-name variable.
+  if (nodeEnv === 'development' && !env.DEV_DB_NAME) {
+    issues.push({
+      path: 'DEV_DB_NAME',
+      message: 'DEV_DB_NAME is required for development',
+      level: 'error',
+    });
+  }
+  if (nodeEnv === 'test' && !env.TEST_DB_NAME) {
+    issues.push({
+      path: 'TEST_DB_NAME',
+      message: 'TEST_DB_NAME is required for test',
+      level: 'error',
+    });
+  }
+  if (nodeEnv === 'production' && !env.PROD_DB_NAME) {
+    issues.push({
+      path: 'PROD_DB_NAME',
+      message: 'PROD_DB_NAME is required for production',
+      level: 'error',
+    });
+  }
+  return issues;
+};
+
+const productionOnlyCheck = (env: EnvMap): ValidationIssue[] => {
+  if (env.NODE_ENV !== 'production') {
+    return [];
+  }
+  const issues: ValidationIssue[] = [];
+
+  // CORS
+  if (!env.CORS_ORIGIN) {
+    issues.push({
+      path: 'CORS_ORIGIN',
+      message: 'CORS_ORIGIN is required in production',
+      level: 'error',
+    });
+  } else {
+    const corsResult = envCorsOriginField.safeParse(env.CORS_ORIGIN);
+    if (!corsResult.success) {
+      for (const err of corsResult.error.issues) {
+        issues.push({ path: 'CORS_ORIGIN', message: err.message, level: 'error' });
+      }
+    }
+  }
+
+  // ADS-410: FRONTEND_URL / RESCUE_FRONTEND_URL must be set in production
+  // so password-reset / verification / invitation emails don't link to localhost.
+  for (const name of ['FRONTEND_URL', 'RESCUE_FRONTEND_URL'] as const) {
+    const value = env[name];
+    if (!value) {
+      issues.push({
+        path: name,
+        message: `${name} is required in production (used to build email links)`,
+        level: 'error',
+      });
+      continue;
+    }
+    const result = envUrlField(name).safeParse(value);
+    if (!result.success) {
+      for (const err of result.error.issues) {
+        issues.push({ path: name, message: err.message, level: 'error' });
+      }
+    }
+  }
+
+  // ADS-542: UPLOAD_SIGNING_SECRET is required in production. The
+  // optionalSecretField above validates length/placeholder when set —
+  // this enforces presence specifically in production.
+  if (!env.UPLOAD_SIGNING_SECRET) {
+    issues.push({
+      path: 'UPLOAD_SIGNING_SECRET',
+      message:
+        'UPLOAD_SIGNING_SECRET is required in production. Generate with: npm run secrets:generate',
+      level: 'error',
+    });
+  }
+
+  // ADS-411: Statsig server secret. Missing -> feature flags silently
+  // default-off. Fail loudly so operators see the misconfiguration.
+  if (!env.STATSIG_SERVER_SECRET_KEY) {
+    issues.push({
+      path: 'STATSIG_SERVER_SECRET_KEY',
+      message:
+        'STATSIG_SERVER_SECRET_KEY is required in production — missing key silently disables ' +
+        'all server-side feature flags',
+      level: 'error',
+    });
+  }
+
+  // ADS-512: DEBUG_ERRORS leaks raw error messages to clients — never allow in prod.
+  if (env.DEBUG_ERRORS === 'true') {
+    issues.push({
+      path: 'DEBUG_ERRORS',
+      message:
+        'DEBUG_ERRORS=true is not allowed in production (leaks raw error messages to clients)',
+      level: 'error',
+    });
+  }
+
+  // BCRYPT_ROUNDS warning in prod
+  const bcryptRounds = env.BCRYPT_ROUNDS;
+  if (bcryptRounds && Number(bcryptRounds) < 12) {
+    issues.push({
+      path: 'BCRYPT_ROUNDS',
+      message: 'BCRYPT_ROUNDS should be at least 12 for production security',
+      level: 'warning',
+    });
+  }
+
+  // DB_LOGGING warning in prod
+  if (env.DB_LOGGING === 'true') {
+    issues.push({
+      path: 'DB_LOGGING',
+      message: 'Database logging is enabled in production - this may log sensitive data',
+      level: 'warning',
+    });
+  }
+
+  return issues;
+};
+
+/**
+ * Validates the supplied env map against the schema and the per-environment
+ * refinements. Pure function so the CLI script and the boot path can both
+ * share it without side effects.
+ */
+export function validateEnv(env: EnvMap): EnvValidationResult {
+  const errors: ValidationIssue[] = [];
+  const warnings: ValidationIssue[] = [];
+
+  const result = envBaseSchema.safeParse(env);
+  if (!result.success) {
+    for (const issue of result.error.issues) {
+      errors.push({
+        path: issue.path.join('.') || '<env>',
+        message: issue.message,
+        level: 'error',
+      });
+    }
+  }
+
+  const collect = (issues: ValidationIssue[]): void => {
+    for (const issue of issues) {
+      if (issue.level === 'error') {
+        errors.push(issue);
+      } else {
+        warnings.push(issue);
+      }
+    }
+  };
+
+  collect(dbNameCheck(env));
+  collect(distinctSecretsCheck(env));
+  collect(productionOnlyCheck(env));
+
+  return { ok: errors.length === 0, errors, warnings };
+}
