+++ package.json
@@ -84,7 +84,7 @@
     "prod:up": "docker compose -f docker-compose.yml -f docker-compose.prod.yml up -d",
     "prod:down": "docker compose -f docker-compose.yml -f docker-compose.prod.yml down",
     "secrets:generate": "node scripts/generate-secrets.mjs",
-    "validate:env": "node scripts/validate-env.mjs",
+    "validate:env": "tsx scripts/validate-env.ts",
     "graph": "turbo run build --graph=docs/dependency-graph.html",
     "graph:tasks": "turbo run build --graph",
     "check:lib-tests": "node scripts/check-lib-tests.mjs",
@@ -116,6 +116,7 @@
     "husky": "^9.1.0",
     "jsdom": "^29.1.1",
     "prettier": "^3.2.5",
+    "tsx": "^4.19.2",
     "turbo": "^2.9.16",
     "typescript": "^6.0.3",
     "typescript-eslint": "^8.60.1",
