--- scripts/validate-env.mjs
@@ -1,425 +1,0 @@
-#!/usr/bin/env node
-
-/**
- * Real environment / secret validator (ADS-408).
- *
- * Replaces the previous URL-only check. This script:
- *  1. Loads the active .env (defaults to repo-root /.env, or --env-file=PATH).
- *  2. Validates required vars per NODE_ENV using the same rules as the
- *     backend startup validator (service.backend/src/utils/validate-env.ts).
- *  3. Diffs the active env against root /.env.example so missing or stale
- *     keys surface before deploy.
- *  4. Exits non-zero on any error so it can run as a CI gate.
- *  5. Optional --staging-env=PATH: fails if any of the six application secrets
- *     share a value with the staging env (prevents staging→prod secret reuse,
- *     ADS-659).
- *
- * Keep the rule set in sync with service.backend/src/utils/validate-env.ts.
- * The two implementations validate the same secrets so operators get the same
- * answer pre-deploy as at boot.
- */
-
-import fs from 'node:fs';
-import path from 'node:path';
-import { fileURLToPath } from 'node:url';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const rootDir = path.dirname(__dirname);
-
-const colors = {
-  red: '\x1b[31m',
-  green: '\x1b[32m',
-  yellow: '\x1b[33m',
-  blue: '\x1b[34m',
-  reset: '\x1b[0m',
-  bold: '\x1b[1m',
-};
-
-const log = {
-  error: msg => console.error(`${colors.red}✗ ${msg}${colors.reset}`),
-  success: msg => console.info(`${colors.green}✓ ${msg}${colors.reset}`),
-  warn: msg => console.warn(`${colors.yellow}! ${msg}${colors.reset}`),
-  info: msg => console.info(`${colors.blue}i ${msg}${colors.reset}`),
-  title: msg => console.info(`${colors.bold}${colors.blue}${msg}${colors.reset}`),
-};
-
-const MIN_SECRET_LENGTH = 32;
-const ENCRYPTION_KEY_HEX_LEN = 64;
-
-// ---------------------------------------------------------------------------
-// .env parsing — minimal KEY=value, ignores blank/comment lines
-// ---------------------------------------------------------------------------
-function parseDotEnv(content) {
-  const out = {};
-  for (const rawLine of content.split('\n')) {
-    const line = rawLine.trim();
-    if (!line || line.startsWith('#')) {
-      continue;
-    }
-    const eq = line.indexOf('=');
-    if (eq === -1) {
-      continue;
-    }
-    const key = line.slice(0, eq).trim();
-    if (!/^[A-Z0-9_]+$/i.test(key)) {
-      continue;
-    }
-    let value = line.slice(eq + 1).trim();
-    if (
-      (value.startsWith('"') && value.endsWith('"')) ||
-      (value.startsWith("'") && value.endsWith("'"))
-    ) {
-      value = value.slice(1, -1);
-    }
-    out[key] = value;
-  }
-  return out;
-}
-
-function loadEnvFile(envPath) {
-  if (!fs.existsSync(envPath)) {
-    return null;
-  }
-  try {
-    return parseDotEnv(fs.readFileSync(envPath, 'utf8'));
-  } catch (err) {
-    log.error(`Failed to read ${envPath}: ${err.message}`);
-    return null;
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Validation rules (mirror service.backend/src/utils/validate-env.ts)
-// ---------------------------------------------------------------------------
-function isPlaceholder(value) {
-  return typeof value === 'string' && value.startsWith('CHANGE_THIS');
-}
-
-function checkSecret(name, value, errors, { required }) {
-  if (!value) {
-    if (required) {
-      errors.push(`${name} is required`);
-    }
-    return;
-  }
-  if (value.length < MIN_SECRET_LENGTH) {
-    errors.push(
-      `${name} must be at least ${MIN_SECRET_LENGTH} characters long for security (got ${value.length})`
-    );
-  }
-  if (isPlaceholder(value)) {
-    errors.push(`${name} must not use the default placeholder value (CHANGE_THIS...)`);
-  }
-}
-
-function checkEncryptionKey(value, errors, { required }) {
-  if (!value) {
-    if (required) {
-      errors.push('ENCRYPTION_KEY is required');
-    }
-    return;
-  }
-  if (value.length !== ENCRYPTION_KEY_HEX_LEN) {
-    errors.push(
-      `ENCRYPTION_KEY must be exactly ${ENCRYPTION_KEY_HEX_LEN} hex characters (32 bytes); got ${value.length}`
-    );
-  } else if (!/^[0-9a-f]+$/i.test(value)) {
-    errors.push('ENCRYPTION_KEY must be hex (0-9, a-f)');
-  }
-}
-
-function checkDistinctSecrets(env, errors) {
-  const pairs = [
-    ['JWT_SECRET', 'JWT_REFRESH_SECRET'],
-    ['JWT_SECRET', 'SESSION_SECRET'],
-    ['JWT_SECRET', 'CSRF_SECRET'],
-    ['JWT_REFRESH_SECRET', 'SESSION_SECRET'],
-    ['JWT_REFRESH_SECRET', 'CSRF_SECRET'],
-    ['SESSION_SECRET', 'CSRF_SECRET'],
-    // ADS-542: UPLOAD_SIGNING_SECRET must be distinct from every other secret.
-    ['UPLOAD_SIGNING_SECRET', 'JWT_SECRET'],
-    ['UPLOAD_SIGNING_SECRET', 'JWT_REFRESH_SECRET'],
-    ['UPLOAD_SIGNING_SECRET', 'SESSION_SECRET'],
-    ['UPLOAD_SIGNING_SECRET', 'CSRF_SECRET'],
-    ['UPLOAD_SIGNING_SECRET', 'ENCRYPTION_KEY'],
-    ['UPLOAD_SIGNING_SECRET', 'JWT_REPORT_SHARE_SECRET'],
-  ];
-  for (const [a, b] of pairs) {
-    if (env[a] && env[b] && env[a] === env[b]) {
-      errors.push(`${a} and ${b} must be distinct (reuse increases compromise blast radius)`);
-    }
-  }
-}
-
-function checkUrl(name, value, errors) {
-  if (!value) {
-    return;
-  }
-  try {
-    new URL(value);
-  } catch {
-    // Don't echo the raw value back — even non-secret env vars can carry
-    // PII or tenant-identifying URLs. The variable name is enough context.
-    errors.push(`${name} must be a valid URL`);
-  }
-}
-
-function validate(env) {
-  const errors = [];
-  const warnings = [];
-  const nodeEnv = env.NODE_ENV || 'development';
-  const isProduction = nodeEnv === 'production';
-
-  if (!['development', 'test', 'production'].includes(nodeEnv)) {
-    errors.push(`NODE_ENV must be one of development|test|production`);
-  }
-
-  // Secrets — required everywhere so a dev config can't leak into prod.
-  checkSecret('JWT_SECRET', env.JWT_SECRET, errors, { required: true });
-  checkSecret('JWT_REFRESH_SECRET', env.JWT_REFRESH_SECRET, errors, { required: true });
-  checkSecret('SESSION_SECRET', env.SESSION_SECRET, errors, { required: true });
-  checkSecret('CSRF_SECRET', env.CSRF_SECRET, errors, { required: true });
-  checkEncryptionKey(env.ENCRYPTION_KEY, errors, { required: true });
-
-  if (env.JWT_REPORT_SHARE_SECRET) {
-    checkSecret('JWT_REPORT_SHARE_SECRET', env.JWT_REPORT_SHARE_SECRET, errors, {
-      required: false,
-    });
-  }
-
-  // ADS-542: UPLOAD_SIGNING_SECRET required in production, optional in dev/test.
-  checkSecret('UPLOAD_SIGNING_SECRET', env.UPLOAD_SIGNING_SECRET, errors, {
-    required: isProduction,
-  });
-
-  checkDistinctSecrets(env, errors);
-
-  // DB connection
-  for (const name of ['DB_HOST', 'DB_PORT', 'DB_USERNAME', 'DB_PASSWORD']) {
-    if (!env[name]) {
-      errors.push(`${name} is required`);
-    }
-  }
-  if (env.DB_PORT && Number.isNaN(Number(env.DB_PORT))) {
-    errors.push(`DB_PORT must be a number (got "${env.DB_PORT}")`);
-  }
-
-  // Per-env DB names (ADS-409/452/465)
-  if (nodeEnv === 'development' && !env.DEV_DB_NAME) {
-    errors.push('DEV_DB_NAME is required for development');
-  }
-  if (nodeEnv === 'test' && !env.TEST_DB_NAME) {
-    errors.push('TEST_DB_NAME is required for test');
-  }
-  if (nodeEnv === 'production' && !env.PROD_DB_NAME) {
-    errors.push('PROD_DB_NAME is required for production');
-  }
-
-  // Booleans
-  for (const name of ['DB_LOGGING', 'WORKER_ENABLED', 'DEBUG_ERRORS']) {
-    if (env[name] && !['true', 'false'].includes(env[name].toLowerCase())) {
-      errors.push(`${name} must be 'true' or 'false' (got "${env[name]}")`);
-    }
-  }
-
-  // ADS-662: validate LOG_LEVEL against Winston's level set
-  const validLogLevels = ['error', 'warn', 'info', 'http', 'debug', 'silly'];
-  if (env.LOG_LEVEL && !validLogLevels.includes(env.LOG_LEVEL.toLowerCase())) {
-    errors.push(
-      `LOG_LEVEL must be one of ${validLogLevels.join('|')} (got "${env.LOG_LEVEL}")`
-    );
-  }
-
-  // Production-only requirements
-  if (isProduction) {
-    if (!env.CORS_ORIGIN) {
-      errors.push('CORS_ORIGIN is required in production');
-    } else if (env.CORS_ORIGIN.split(',').some(o => o.trim() === '*' || o.includes('*'))) {
-      errors.push("CORS_ORIGIN cannot contain wildcard ('*') in production");
-    }
-    // ADS-668: block dev-style origins in production CORS
-    if (env.CORS_ORIGIN) {
-      const devPatterns = ['localhost', '127.0.0.1', '0.0.0.0', '.ngrok.io', '.ngrok-free.app'];
-      const origins = env.CORS_ORIGIN.split(',').map(o => o.trim());
-      const devOrigins = origins.filter(o =>
-        devPatterns.some(p => o.includes(p))
-      );
-      if (devOrigins.length > 0) {
-        errors.push(
-          `CORS_ORIGIN contains dev-style origins in production: ${devOrigins.join(', ')}. ` +
-          'Use real domain names only'
-        );
-      }
-      const httpOrigins = origins.filter(o => o.startsWith('http://'));
-      if (httpOrigins.length > 0) {
-        warnings.push(
-          `CORS_ORIGIN contains non-HTTPS origins: ${httpOrigins.join(', ')}. ` +
-          'Production should use HTTPS only'
-        );
-      }
-    }
-    // ADS-410
-    if (!env.FRONTEND_URL) {
-      errors.push('FRONTEND_URL is required in production (used to build email links)');
-    } else {
-      checkUrl('FRONTEND_URL', env.FRONTEND_URL, errors);
-    }
-    if (!env.RESCUE_FRONTEND_URL) {
-      errors.push('RESCUE_FRONTEND_URL is required in production (used to build email links)');
-    } else {
-      checkUrl('RESCUE_FRONTEND_URL', env.RESCUE_FRONTEND_URL, errors);
-    }
-    // ADS-411
-    if (!env.STATSIG_SERVER_SECRET_KEY) {
-      errors.push(
-        'STATSIG_SERVER_SECRET_KEY is required in production (missing key silently disables all server-side feature flags)'
-      );
-    }
-    // ADS-512
-    if (env.DEBUG_ERRORS === 'true') {
-      errors.push(
-        'DEBUG_ERRORS=true is not allowed in production (leaks raw error messages to clients)'
-      );
-    }
-    // BCRYPT_ROUNDS warning
-    if (env.BCRYPT_ROUNDS && Number(env.BCRYPT_ROUNDS) < 12) {
-      warnings.push('BCRYPT_ROUNDS should be at least 12 for production security');
-    }
-    if (env.DB_LOGGING === 'true') {
-      warnings.push('DB_LOGGING is enabled in production (may log sensitive data)');
-    }
-  }
-
-  return { errors, warnings };
-}
-
-// ---------------------------------------------------------------------------
-// Diff: keys present in .env.example but missing from active env
-// ---------------------------------------------------------------------------
-function diffAgainstExample(env, examplePath) {
-  if (!fs.existsSync(examplePath)) {
-    return { missing: [], examplePath: null };
-  }
-  const example = parseDotEnv(fs.readFileSync(examplePath, 'utf8'));
-  const missing = Object.keys(example).filter(key => !(key in env));
-  return { missing, examplePath };
-}
-
-// ---------------------------------------------------------------------------
-// Main
-// ---------------------------------------------------------------------------
-// Staging-reuse check (ADS-659)
-// Fails if any of the six application secrets share a value with a staging env.
-// ---------------------------------------------------------------------------
-const SECRET_VARS = [
-  'JWT_SECRET',
-  'JWT_REFRESH_SECRET',
-  'SESSION_SECRET',
-  'CSRF_SECRET',
-  'ENCRYPTION_KEY',
-  'UPLOAD_SIGNING_SECRET',
-];
-
-function checkStagingReuse(prodEnv, stagingPath, errors) {
-  const stagingEnv = loadEnvFile(stagingPath);
-  if (!stagingEnv) {
-    log.warn(`--staging-env file not found at ${stagingPath}; skipping reuse check`);
-    return;
-  }
-  for (const key of SECRET_VARS) {
-    const prodVal = prodEnv[key];
-    const stagingVal = stagingEnv[key];
-    if (prodVal && stagingVal && prodVal === stagingVal) {
-      errors.push(
-        `${key} is identical in production and staging — rotate before deploying`
-      );
-    }
-  }
-}
-
-function parseArgs(argv) {
-  const args = { envFile: null, examplePath: null, stagingEnv: null, json: false };
-  for (const arg of argv.slice(2)) {
-    if (arg.startsWith('--env-file=')) {
-      args.envFile = arg.slice('--env-file='.length);
-    } else if (arg.startsWith('--example=')) {
-      args.examplePath = arg.slice('--example='.length);
-    } else if (arg.startsWith('--staging-env=')) {
-      args.stagingEnv = arg.slice('--staging-env='.length);
-    } else if (arg === '--json') {
-      args.json = true;
-    }
-  }
-  return args;
-}
-
-function main() {
-  const args = parseArgs(process.argv);
-  const envFile = args.envFile ? path.resolve(args.envFile) : path.join(rootDir, '.env');
-  const examplePath = args.examplePath
-    ? path.resolve(args.examplePath)
-    : path.join(rootDir, '.env.example');
-  log.title('Environment / secret validation');
-  log.info(`env file:    ${envFile}`);
-  log.info(`example:     ${examplePath}`);
-  if (args.stagingEnv) {
-    log.info(`staging env: ${path.resolve(args.stagingEnv)}`);
-  }
-
-  const fileEnv = loadEnvFile(envFile);
-  // Merge file env on top of process.env so `STATSIG_SERVER_SECRET_KEY=foo
-  // npm run validate:env` works for CI. process.env wins for already-set
-  // overrides (CI typically injects secrets that way).
-  const env = { ...(fileEnv || {}), ...process.env };
-
-  if (!fileEnv) {
-    log.warn(`No .env file at ${envFile}; validating process.env only`);
-  }
-
-  const { errors, warnings } = validate(env);
-
-  if (args.stagingEnv) {
-    checkStagingReuse(env, path.resolve(args.stagingEnv), errors);
-  }
-  const { missing, examplePath: foundExample } = diffAgainstExample(env, examplePath);
-
-  if (foundExample && missing.length > 0) {
-    log.warn(`${missing.length} key(s) in .env.example missing from active env:`);
-    for (const key of missing) {
-      console.info(`    - ${key}`);
-    }
-  } else if (foundExample) {
-    log.success('All keys from .env.example are present in active env');
-  }
-
-  if (warnings.length > 0) {
-    log.warn(`${warnings.length} warning(s):`);
-    for (const w of warnings) {
-      console.warn(`    - ${w}`);
-    }
-  }
-
-  if (args.json) {
-    console.info(
-      JSON.stringify(
-        { ok: errors.length === 0, errors, warnings, missingFromExample: missing },
-        null,
-        2
-      )
-    );
-  }
-
-  if (errors.length > 0) {
-    log.error(`Validation failed with ${errors.length} error(s):`);
-    for (const e of errors) {
-      console.error(`    - ${e}`);
-    }
-    process.exit(1);
-  }
-
-  log.success('Environment validation passed');
-  process.exit(0);
-}
-
-main();
